smartbus-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
firebase-admin==6.2.0
websockets==12.0
httpx==0.25.2
pydantic==2.5.0
numpy==1.26.2
pytest==7.4.3
pytest-asyncio==0.21.1
